+++ src/components/semantic/presenters/FigurePresenter.tsx
@@ -13,9 +13,9 @@ import { useFixedRef } from "../../../utils/hooks";
 import styles from "../styles/figure.module.css";
 import {NON_STATIC_FIGURE_FLAG} from "../../../isaac/IsaacTypes";
 import {Alert, Input, Label} from "reactstrap";
-import { DropZoneQuestionContext } from "./ItemQuestionPresenter";
+import { DropZoneQuestionContext } from "./questions/ItemQuestionPresenter";
 import { FigureRegionModal } from "../../FigureRegionModal";
-import { InlineQuestionContext } from "./questionPresenters";
+import { InlineQuestionContext } from "./questions/questionPresenters";
 import { isDefined } from "../../../utils/types";
 
 function githubURLFromGithubData(data: {download_url: string}, svgView?: string | null) {
